requirements.txt
astroid==1.6.1
click==6.7
colorama==0.3.9
enum34==1.1.6
eventlet==0.17.4
Flask==0.12.2
Flask-SocketIO==2.9.3
greenlet==0.4.13
gunicorn==19.7.1
isort==4.2.15
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
numpy==1.14.0
Pillow==5.0.0
pylint==1.8.2
python-engineio==2.0.2
python-socketio==1.8.4
six==1.11.0
Werkzeug==0.14.1
wrapt==1.10.11
rauth==0.7.3
opencv-python
